refactor(c1-b1): split cli args mod behind thin facade
