+++ crates/assay-cli/src/cli/args/evidence.rs
@@ -1,0 +1,5 @@
+#[derive(clap::Args, Clone, Debug)]
+pub struct EvidenceArgs {
+    #[command(subcommand)]
+    pub cmd: crate::cli::commands::evidence::EvidenceCmd,
+}
